fix: flow registration payment ganda dewasa & fixing landing page

+++ database/migrations/2026_03_11_122851_remove_alamat_from_registrations_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('registrations', function (Blueprint $table) {
-            $table->dropColumn('alamat');
-        });
+        if (Schema::hasColumn('registrations', 'alamat')) {
+            Schema::table('registrations', function (Blueprint $table) {
+                $table->dropColumn('alamat');
+            });
+        }
     }
 
     public function down(): void
